Separate out raw2trace code shared with tracer
Extracts raw2trace code that is shared with the tracer into a separate
raw2trace_shared.h and raw2trace_shared.cpp. This is to allow us to
skip including raw2trace libraries if building drmemtrace only for
tracing.
